Backend: Cleaning dogs and users controllers
